+++ previous-versions/files/src/sdk/models/operations/downloadfiles.ts
@@ -97,7 +97,7 @@ export class DownloadFilesResponse extends SpeakeasyBase {
      * Raw HTTP response; suitable for custom response parsing
      */
     @SpeakeasyMetadata()
-    rawResponse?: AxiosResponse;
+    rawResponse: AxiosResponse;
 
     /**
      * The request made is not valid.