+++ previous-versions/files/src/sdk/models/operations/listfiles.ts
@@ -91,7 +91,7 @@ export class ListFilesResponse extends SpeakeasyBase {
      * Raw HTTP response; suitable for custom response parsing
      */
     @SpeakeasyMetadata()
-    rawResponse?: AxiosResponse;
+    rawResponse: AxiosResponse;
 
     /**
      * Your API request was not properly authorized.