+++ previous-versions/files/src/sdk/models/operations/uploadfiles.ts
@@ -104,7 +104,7 @@ export class UploadFilesResponse extends SpeakeasyBase {
      * Raw HTTP response; suitable for custom response parsing
      */
     @SpeakeasyMetadata()
-    rawResponse?: AxiosResponse;
+    rawResponse: AxiosResponse;
 
     /**
      * The request made is not valid.