+++ previous-versions/files/src/sdk/sdk.ts
@@ -54,9 +54,9 @@ export class SDKConfiguration {
     serverDefaults: any;
     language = "typescript";
     openapiDocVersion = "3.0.0";
-    sdkVersion = "0.38.0";
-    genVersion = "2.188.1";
-    userAgent = "speakeasy-sdk/typescript 0.38.0 2.188.1 3.0.0 @codat/files";
+    sdkVersion = "0.38.1";
+    genVersion = "2.194.1";
+    userAgent = "speakeasy-sdk/typescript 0.38.1 2.194.1 3.0.0 @codat/files";
     retryConfig?: utils.RetryConfig;
     public constructor(init?: Partial<SDKConfiguration>) {
         Object.assign(this, init);