+++ previous-versions/files/RELEASES.md
@@ -617,3 +617,13 @@ Based on:
 - [typescript v0.38.0] previous-versions/files
 ### Releases
 - [NPM v0.38.0] https://www.npmjs.com/package/@codat/files/v/0.38.0 - previous-versions/files
+
+## 2023-11-20 01:29:07
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Files.yaml
+- Speakeasy CLI 1.121.1 (2.194.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.38.1] previous-versions/files
+### Releases
+- [NPM v0.38.1] https://www.npmjs.com/package/@codat/files/v/0.38.1 - previous-versions/files